Fix #507: Moved debugging functions from cx/utilities.go to cx/debug.go.
